refactor: move constants file to constants folder
