--- src/constants.ts
@@ -1,10 +1,0 @@
-export const DEFAULT_PORT = 4000;
-
-export const YAML_SWAGGER_FILENAME = 'api.yaml';
-
-export const PATH_TO_YAML_FOLDER = './doc';
-
-export enum Messages {
-  OLD_PASSWORD_WRONG = 'Old password is wrong',
-  ARTIST_OR_ALBUM_DOES_NOT_EXIST = 'Artist and/or album with provided id does not exist',
-}
